+++ CHANGES.md
@@ -6,6 +6,8 @@ Olza Logistic's Pickup Point API Client PHP library.
 
 * dev
   * Changed `Country::CZECH_REPUBLIC` to `Country::CZECHIA`.
+  * Added support for "payments" and "services" find() args.
+  *
   * Updated dependencies.
 
 * v1.0.0 (2022-02-16)